+++ app/src/main/java/com/example/swapit/data/datasource/remote/dto/response/user/ReviewResponse.kt
@@ -1,0 +1,13 @@
+package com.example.swapit.data.datasource.remote.dto.response.user
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ReviewResponse(
+    val userId: Long,
+    val profileImageUrl: String,
+    val nickName: String,
+    val rating: Double,
+    val content: String,
+    val createdAt: String,
+)
